requirements.txt
requests>=2.28.0
PySocks>=1.7.0
python-dotenv>=1.0.0
rich>=13.0.0
beautifulsoup4>=4.12.0
geoip2>=4.8.0
